% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{An object belonging to the Turbidata class}
\format{
data$data
\describe{
\item{height}{height, in cm}
\item{time}{time, in min}
\item{signal}{signal, can be T or R}
\item{do}{the value}
}
}
\usage{
data
}
\description{
An object belonging to the Turbidata class
}
\keyword{datasets}
