% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_turbidata.R
\name{relative_turbidata}
\alias{relative_turbidata}
\title{TurbiRelativeData}
\description{
R6 Class representing the data convert in relative value.
}
\section{Warning}{

Do not use $new() of this class directly : use \link[=turbidata]{TurbiDumpData$use_relative()}
}

\section{Super class}{
\code{\link[turbidata:TurbiDumpData]{turbidata::TurbiDumpData}} -> \code{TurbiRelativeData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TurbiRelativeData-new}{\code{relative_turbidata$new()}}
\item \href{#method-TurbiRelativeData-clone}{\code{relative_turbidata$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="turbidata" data-topic="TurbiDumpData" data-id="finalize"><a href='../../turbidata/html/TurbiDumpData.html#method-TurbiDumpData-finalize'><code>turbidata::TurbiDumpData$finalize()</code></a></li>
<li><span class="pkg-link" data-pkg="turbidata" data-topic="TurbiDumpData" data-id="use_absolute"><a href='../../turbidata/html/TurbiDumpData.html#method-TurbiDumpData-use_absolute'><code>turbidata::TurbiDumpData$use_absolute()</code></a></li>
<li><span class="pkg-link" data-pkg="turbidata" data-topic="TurbiDumpData" data-id="use_relative"><a href='../../turbidata/html/TurbiDumpData.html#method-TurbiDumpData-use_relative'><code>turbidata::TurbiDumpData$use_relative()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiRelativeData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiRelativeData-new}{}}}
\subsection{Method \code{new()}}{
Create a new TurbiRelativeData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{relative_turbidata$new(data, signal, name, dump_data_fille)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{data.frame} the relative data.frame}

\item{\code{signal}}{\code{vector} A vector containing what value are contain in data.frame}

\item{\code{name}}{\code{character} The name of the sample}

\item{\code{dump_data_fille}}{\code{character} The path of tmp dump data file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TurbiRelativeData} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiRelativeData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiRelativeData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{relative_turbidata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
