% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidata.R
\name{turbidata}
\alias{turbidata}
\title{TurbiDumpData}
\description{
R6 Class representing the raw values of the turbiscan
initialize
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{\code{character} The name of the sample}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TurbiDumpData-new}{\code{turbidata$new()}}
\item \href{#method-TurbiDumpData-use_relative}{\code{turbidata$use_relative()}}
\item \href{#method-TurbiDumpData-use_absolute}{\code{turbidata$use_absolute()}}
\item \href{#method-TurbiDumpData-clone}{\code{turbidata$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiDumpData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiDumpData-new}{}}}
\subsection{Method \code{new()}}{
Create a new TurbiDumpData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata$new(
  file,
  dir = "data/",
  ext = "auto",
  cache_dir = "cache/",
  force_update_cache = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{\code{character}}

\item{\code{dir}}{\code{character} (optional) The folder containing the data file (default \code{"data/"})}

\item{\code{ext}}{\code{character} (optional) The data format. Value available : 'auto', .csv' and '.xls' (or 'xlsx' alias) (default \code{'csv'})}

\item{\code{cache_dir}}{\code{character} (optional) The folder containing the cache file (default \code{"cache/"})}

\item{\code{force_update_cache}}{\code{logical} (optional) Should we ignore the cache value and use the data file (default \code{FALSE})}

\item{\code{create_cache_file}}{\code{logical} (optional) Should we create cache file (default \code{TRUE})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TurbiDumpData} object.
use_relative
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiDumpData-use_relative"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiDumpData-use_relative}{}}}
\subsection{Method \code{use_relative()}}{
Allows you to calculate on the relative values of the data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata$use_relative(r = 1, t = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{(optional) the time used as reference for the backscatter (default: \code{1})}

\item{\code{t}}{(optional) the time used as reference for the transmission (default: \code{1})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link[=relative_turbidata]{TurbiRelativeData} object.
use_absolute
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiDumpData-use_absolute"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiDumpData-use_absolute}{}}}
\subsection{Method \code{use_absolute()}}{
Allows you to calculate on the absolute values of the data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata$use_absolute(r = 1, t = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{(optional) the time used as reference for the backscatter (default: \code{1})}

\item{\code{t}}{(optional) the time used as reference for the transmission (default: \code{1})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \link[=relative_turbidata]{TurbiAbsoluteData} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiDumpData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiDumpData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
