% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidata_class.R
\name{turbidata_class}
\alias{turbidata_class}
\title{Turbidata class}
\description{
Turbidata class

Turbidata class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{The \code{data.frame} (read only)}

\item{\code{dump_data}}{\code{data.frame} The dump_data (read only)}

\item{\code{name}}{\code{character} The name of the sample}

\item{\code{type}}{\code{character} Can be \code{"absolute"} or \code{"relative"} : the type of actual data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TurbiData-new}{\code{turbidata_class$new()}}
\item \href{#method-TurbiData-print}{\code{turbidata_class$print()}}
\item \href{#method-TurbiData-use_absolute}{\code{turbidata_class$use_absolute()}}
\item \href{#method-TurbiData-use_relative}{\code{turbidata_class$use_relative()}}
\item \href{#method-TurbiData-average_area}{\code{turbidata_class$average_area()}}
\item \href{#method-TurbiData-ggplot}{\code{turbidata_class$ggplot()}}
\item \href{#method-TurbiData-gganim}{\code{turbidata_class$gganim()}}
\item \href{#method-TurbiData-export}{\code{turbidata_class$export()}}
\item \href{#method-TurbiData-clone}{\code{turbidata_class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-new"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-new}{}}}
\subsection{Method \code{new()}}{
Create a new TurbiData object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$new(
  path,
  ext = "auto",
  cache_dir = private$shared$cache_dir,
  force_update_cache = FALSE,
  create_cache_file = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{\code{character} The path to the file}

\item{\code{ext}}{\code{character} (optional) The data format. Value available : 'auto', .csv' and '.xls' (or 'xlsx' alias)}

\item{\code{cache_dir}}{\code{character} (optional) The folder containing the cache file}

\item{\code{force_update_cache}}{\code{logical} (optional) Should we ignore the cache value and use the data file}

\item{\code{create_cache_file}}{\code{logical} (optional) Should we create cache file to load faster next time}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{TurbiData} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-print"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-print}{}}}
\subsection{Method \code{print()}}{
The function use to \code{print()} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-use_absolute"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-use_absolute}{}}}
\subsection{Method \code{use_absolute()}}{
Allows you to calculate on the absolute values of the data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$use_absolute()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-use_relative"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-use_relative}{}}}
\subsection{Method \code{use_relative()}}{
Allows you to calculate on the relative values of the data.frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$use_relative(r = 1, t = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{(optional) the time used as reference for the backscatter}

\item{\code{t}}{(optional) the time used as reference for the transmission}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-average_area"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-average_area}{}}}
\subsection{Method \code{average_area()}}{
This function makes it possible to calculate the area of the average values located between \code{min} and \code{max}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$average_area(
  mini = min(private$..data["height"]),
  maxi = max(private$..data["height"])
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mini}}{\code{number} (optional) The minimum height}

\item{\code{maxi}}{\code{number} (optional) The maximum height}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A clone of self with the clean data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-ggplot"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-ggplot}{}}}
\subsection{Method \code{ggplot()}}{
Make plot with data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$ggplot(
  title = NULL,
  x = "Height (cm)",
  y = "\%",
  color = "Time (min)",
  y_lim = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{\code{character} (optional) The name of plot}

\item{\code{x}}{\code{character} (optional) The labs for x axis}

\item{\code{y}}{\code{character} (optional) The labs for y axis}

\item{\code{color}}{\code{character} (optional) The labs for color gradient}

\item{\code{y_lim}}{\code{vector} (optional) A vector containing min and max value of y limit axe. If missing, this value is auto calculate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-gganim"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-gganim}{}}}
\subsection{Method \code{gganim()}}{
Make annimated plot with data and export to new file. Warning, rendering step may be realy long
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$gganim(
  title = NULL,
  x = "Height (cm)",
  y = "\%",
  color = "Time (min)",
  y_lim = NULL,
  width = 1600L,
  height = 600L,
  fps = 5L,
  export_format = "mp4",
  filename = NULL,
  output_dir = private$shared$output_dir,
  override_file = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{\code{character} (optional) The name of plot}

\item{\code{x}}{\code{character} (optional) The labs for x axis}

\item{\code{y}}{\code{character} (optional) The labs for y axis}

\item{\code{color}}{\code{character} (optional) The labs for color gradient}

\item{\code{y_lim}}{\code{vector} (optional) A vector containing min and max value of y limit axe. If missing, this value is auto calculate.}

\item{\code{width}}{\code{integer} (optional): With of the render}

\item{\code{height}}{\code{integer} (optional): Height of the render}

\item{\code{fps}}{\code{integer} (optional): Number of frame by second}

\item{\code{export_format}}{\code{integer} (optional) The output format. May be 'mp4' or 'gif'}

\item{\code{filename}}{\code{character} (optional) The filename use for export. Default : same as data file}

\item{\code{output_dir}}{\code{integer} (optional) The output directory}

\item{\code{override_file}}{\code{logical} (optional) Override file if already exist}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-export"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-export}{}}}
\subsection{Method \code{export()}}{
Export data.frame to file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$export(
  format = "csv",
  output_dir = private$shared$output_dir,
  override_file = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{(optional) The output format}

\item{\code{output_dir}}{(optional) The output directory}

\item{\code{override_file}}{(optional) Override file if already exist}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TurbiData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TurbiData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{turbidata_class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
