% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbidata.R, R/turbidata_function.R
\docType{package}
\name{turbidata}
\alias{turbidata}
\title{turbidata:}
\usage{
turbidata(...)
}
\arguments{
\item{...}{Argument transfert to turbidata_class$new()}
}
\description{
A package to rapid and simplified analysis of data from the turbiscan CLASSIC from Formulaction

turbidata() : Wrapper for \link[=turbidata_class]{Turbidata class}
}
